Move setAdSlotMinHeight function to on-slot-viewable file. Call function before retrieval of non-refreshable line items
